Removes unused CSS on build
Now using PostCSS plugin PurgeCSS to get rid of unused *inline* styles. Wasn't easy out of the box, but found a handy example of doing this with some JS magic. A little wonky using PostCSS in a script *and* using it again in my eleventy config, but that's the way the cookie crumbles. 🍪